+++ app/src/main/java/pt/rvcoding/cvnotes/domain/model/SectionWithNotes.kt
@@ -1,5 +1,8 @@
 package pt.rvcoding.cvnotes.domain.model
 
+import androidx.compose.runtime.Immutable
+
+@Immutable
 data class SectionWithNotes(
     val section: Section,
     val notes: List<Note>